<BRMS-2852> Move the ImportModal to proper module.
