+++ examples/editor/example.js
@@ -4,12 +4,9 @@
 import React from 'react';
 import DeckGL from '@deck.gl/react';
 import { EditableGeoJsonLayer } from '@nebula.gl/layers';
-import { Toolbox } from '@nebula.gl/editor';
+import { ImportModal, Toolbox } from '@nebula.gl/editor';
 import { DrawPolygonMode } from '@nebula.gl/edit-modes';
 import { StaticMap } from 'react-map-gl';
-
-import { ModalProvider } from 'styled-react-modal';
-import { FancyModalButton } from './import-modal.js';
 
 const MAPBOX_ACCESS_TOKEN =
   'pk.eyJ1IjoiZ2Vvcmdpb3MtdWJlciIsImEiOiJjanZidTZzczAwajMxNGVwOGZrd2E5NG90In0.gdsRu_UeU_uPi9IulBruXA';
@@ -59,9 +56,7 @@ export function Example() {
         modeConfig={modeConfig}
         onSetModeConfig={setModeConfig}
       />
-      <ModalProvider>
-        <FancyModalButton onImport={geojson => setFeatures(geojson)} />
-      </ModalProvider>
+      <ImportModal onImport={geojson => setFeatures(geojson)} />
     </>
   );
 }
