+++ examples/editor/package.json
@@ -9,11 +9,8 @@
     "@deck.gl/mesh-layers": "^8.0.6",
     "@deck.gl/react": "^8.0.6",
     "react": "^16.8.0",
-    "react-autosize-textarea": "^7.0.0",
     "react-dom": "^16.4.2",
-    "react-map-gl": "^5.2.1",
-    "styled-components": "^4.3.2",
-    "styled-react-modal": "1.2.4"
+    "react-map-gl": "^5.2.1"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
