+++ modules/editor/package.json
@@ -50,6 +50,9 @@
     "@maphubs/tokml": "^0.5.2",
     "@tmcw/togeojson": "^3.2.0",
     "react": "^16.8.0",
+    "react-autosize-textarea": "^7.0.0",
+    "styled-components": "^4.3.2",
+    "styled-react-modal": "1.2.4",
     "terraformer-wkt-parser": "^1.2.1"
   },
   "devDependencies": {
