+++ modules/editor/src/index.js
@@ -1,3 +1,4 @@
 // @flow
 
 export { Toolbox } from './toolbox.js';
+export { ImportModal } from './import-modal.js';
